Refactor persistence controller helpers
